testing region finding, added experimental (currently disabled) reduction of loop gain based of compactness. Less compact selected pixels reduce loop gain. Seems to help with speckling, but increases runtime

+++ deconv_testing/str_printf.cpp
@@ -3,7 +3,7 @@
 // SPECIFIC TYPE OUTPUT
 std::ostream& operator<<(std::ostream& os, const std::byte& a){
 	char buf[5];
-	snprintf(buf, 5, "%x", a);
+	snprintf(buf, 5, "%x", (uint8_t)a);
 	os << buf;
 	return os;
 }
